refactor: extract GitHub status fetching to status.go
Extract the PR/CI status fetching logic from pr_operations.go into a
dedicated status.go file. This makes the code easier to find and
provides a cleaner foundation for the shippable feature.

Changes:
- Create status.go with BatchGetPRChecksStatusGraphQL and helper functions
- Add review decision constants (ReviewDecisionApproved, etc.)
- Add convenience methods on CheckStatus (IsApproved, IsReady, etc.)
- Remove duplicated code from pr_operations.go
- Add status_test.go with tests for the new helper methods

The existing API (Client interface) remains unchanged as it already
delegates to these functions.
